Install package locally
Prior to this change, the contributor instructions did not recommend
installing the first-party `timezone_tools` package locally. This meant
you had to pass a `PYTHONPATH` environment variable to pytest to run
tests.

This isn't a problem when running tests in `tox` as it automatically
installs the first-party package as part of its set-up phase.

This change updates the `makefile` to install the `timezone_tools` in
editable mode, making it possible to run `pytest` (and other tools) from
repo root.